api/requirements.txt
fastapi
uvicorn
databases[postgresql]
asyncpg
pytest
httpx
pytest-asyncio
python-jose[cryptography]
passlib[bcrypt]